+++ src/ext/queries/index.ts
@@ -1,0 +1,7 @@
+import { Config } from './config';
+import { Files } from './files';
+
+export default {
+    config: new Config(),
+    files: new Files(),
+};