+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
+## 2.0.0 (2018-12-08)
+* Bug fixes:
+  * Does not run other editor actions on save [(#5)](https://github.com/jbockle/format-files/issues/5) [(#2)](https://github.com/jbockle/format-files/issues/2)
+* Enhancements:
+  * Add Progress Bar [(#3)](https://github.com/jbockle/format-files/issues/3)
+  * Make cancel formatting more accessible [(#6)](https://github.com/jbockle/format-files/issues/6)
+  * Inherit workspace setting `file.exclude` [(#1)](https://github.com/jbockle/format-files/issues/1)
 ## 1.0.1 (2018-01-02)
 * Updated readme
 ## 1.0.0 (2017-12-29)